refact/move_methods_related_with_refreshToken_to_refreshTokenService#20
