+++ src/main/java/com/habitpay/habitpay/domain/refreshToken/application/NewTokenService.java
@@ -22,12 +22,14 @@ public class NewTokenService {
     private final RefreshTokenService refreshTokenService;
     private final MemberService memberService;
 
-    public String createNewAccessToken(String refreshToken, String requestIp) {
+    public String createNewAccessToken(String refreshToken) {
 
         if (!tokenProvider.validateToken(refreshToken)) {
             throw new IllegalArgumentException("새로운 토큰을 발급해주려 했으나 리프레시 토큰이 이상하네요.");
         }
 
+        String requestIp = refreshTokenService.getClientIpAddress();
+        log.info("Client IP Address : {}", requestIp);
         String loginIp = refreshTokenService.findByRefreshToken(refreshToken).getLoginIp();
         if (!Objects.equals(requestIp, loginIp)) {
             throw new IllegalArgumentException("로그인한 주소와 요청한 주소가 달라요.");
@@ -38,36 +40,5 @@ public class NewTokenService {
 
         // todo : 토큰 유효 기간
         return tokenProvider.generateToken(member, Duration.ofHours(2));
-    }
-
-    public String getClientIpAddress() {
-
-        if (Objects.isNull(RequestContextHolder.getRequestAttributes())) {
-            return "0.0.0.0";
-        }
-
-        final String[] IpHeaderCandidates = {
-                "X-Forwarded-For",
-                "Proxy-Client-IP",
-                "WL-Proxy-Client-IP",
-                "HTTP_X_FORWARDED_FOR",
-                "HTTP_X_FORWARDED",
-                "HTTP_X_CLUSTER_CLIENT_IP",
-                "HTTP_CLIENT_IP",
-                "HTTP_FORWARDED_FOR",
-                "HTTP_FORWARDED",
-                "HTTP_VIA",
-                "REMOTE_ADDR"
-        };
-
-        HttpServletRequest request = ((ServletRequestAttributes) RequestContextHolder.getRequestAttributes()).getRequest();
-        for (String header : IpHeaderCandidates) {
-            String requestIp = request.getHeader(header);
-            if (Objects.nonNull(requestIp) && !requestIp.isEmpty() && !"unknown".equalsIgnoreCase(requestIp)) {
-                String ip = requestIp.split(",")[0];
-                return ip;
-            }
-        }
-        return request.getRemoteAddr();
     }
 }
