+++ src/main/java/com/habitpay/habitpay/global/config/jwt/TokenService.java
@@ -30,7 +30,7 @@ public class TokenService {
         Member member = optionalMember.get();
 
 //        return tokenProvider.generateToken(member, Duration.ofHours(2));
-        return tokenProvider.generateToken(member, Duration.ofMinutes(1));
+        return tokenProvider.generateToken(member, Duration.ofSeconds(30));
     }
 
     public String createRefreshToken(String email) {
